package/requirements.txt
torch>=2.5.1                    
torchvision>=0.20.1 

numpy
opencv-python      
av  

matplotlib
ultralytics
hydra-core